+++ src/enums.rs
@@ -2,7 +2,7 @@ use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
 
 #[pyclass(from_py_object, module = "tibs")]
-#[derive(Clone, Copy, PartialEq, Eq)]
+#[derive(Clone, Copy, PartialEq, Eq, Hash)]
 pub enum Endianness {
     Unspecified,
     Big,
@@ -67,7 +67,7 @@ pub enum Codec {
 }
 
 #[pyclass(from_py_object, module = "tibs")]
-#[derive(Clone, Copy)]
+#[derive(Clone, Copy, PartialEq, Eq, Hash)]
 pub enum DtypeKind {
     Uint,
     Int,